size_t -> unsigned for dl4mt

+++ src/amun/common/sentence.cpp
@@ -7,7 +7,7 @@ using namespace std;
 
 namespace amunmt {
 
-Sentence::Sentence(const God &god, size_t vLineNum, const std::string& line)
+Sentence::Sentence(const God &god, unsigned vLineNum, const std::string& line)
   : lineNum_(vLineNum)
 {
   std::vector<std::string> tabs;
@@ -16,8 +16,8 @@ Sentence::Sentence(const God &god, size_t vLineNum, const std::string& line)
     tabs.push_back("");
   }
 
-  size_t maxLength = god.Get<size_t>("max-length");
-  size_t i = 0;
+  unsigned maxLength = god.Get<unsigned>("max-length");
+  unsigned i = 0;
   for (auto& tab : tabs) {
     std::vector<std::string> lineTokens;
     Trim(tab);
@@ -37,7 +37,7 @@ Sentence::Sentence(const God &god, size_t vLineNum, const std::string& line)
     auto processed = god.Preprocess(i, lineFactors);
     factors_.emplace_back(god.GetSourceVocabs(i)(processed));
     Words lineWords(factors_.back().size());
-    for (size_t i = 0; i < factors_.back().size(); ++i) {
+    for (unsigned i = 0; i < factors_.back().size(); ++i) {
       lineWords[i] = factors_.back()[i][0];
     }
     words_.emplace_back(lineWords);
@@ -45,7 +45,7 @@ Sentence::Sentence(const God &god, size_t vLineNum, const std::string& line)
   }
 }
 
-Sentence::Sentence(const God &god, size_t lineNum, const std::vector<std::string>& words)
+Sentence::Sentence(const God &god, unsigned lineNum, const std::vector<std::string>& words)
   : lineNum_(lineNum) {
     auto processed = god.Preprocess(0, words);
     words_.push_back(god.GetSourceVocab(0)(processed));
@@ -54,7 +54,7 @@ Sentence::Sentence(const God &god, size_t lineNum, const std::vector<std::string
     FillDummyFactors(words_.back());
 }
 
-Sentence::Sentence(God&, size_t lineNum, const std::vector<uint>& words)
+Sentence::Sentence(God&, unsigned lineNum, const std::vector<uint>& words)
   : lineNum_(lineNum) {
     words_.push_back(words);
     // fill in the factors as well so that there aren't any surprises
@@ -65,24 +65,24 @@ Sentence::Sentence(God&, size_t lineNum, const std::vector<uint>& words)
 void Sentence::FillDummyFactors(const Words& line) {
   factors_.emplace_back(FactWords(line.size(), FactWord(1)));
   FactWords& factline = factors_.back();
-  for (size_t i = 0; i < line.size(); ++i) {
+  for (unsigned i = 0; i < line.size(); ++i) {
     factline[i][0] = line[i];
   }
 }
 
-size_t Sentence::GetLineNum() const {
+unsigned Sentence::GetLineNum() const {
   return lineNum_;
 }
 
-const Words& Sentence::GetWords(size_t index) const {
+const Words& Sentence::GetWords(unsigned index) const {
   return words_[index];
 }
 
-const FactWords& Sentence::GetFactors(size_t index) const {
+const FactWords& Sentence::GetFactors(unsigned index) const {
   return factors_[index];
 }
 
-size_t Sentence::size(size_t index) const {
+unsigned Sentence::size(unsigned index) const {
   return words_[index].size();
 }
